CHG: Accuracy improvments
git-svn-id: https://elves.srl.caltech.edu/svn/nusim/trunk@460 f61b1093-7add-4ab6-ac1b-eb05caf01ebb

+++ src/modules/inc/NModuleInterfaceEventSaverLevel2Fits.h
@@ -106,8 +106,8 @@ class NModuleInterfaceEventSaverLevel2Fits : public NModuleInterfaceObservation
   NExposureMap m_ExposureMap;
   
   //! fields for fits file.
-  vector<float>  m_Ra; 
-  vector<float>  m_Dec;
+  vector<double>  m_Ra; 
+  vector<double>  m_Dec;
   vector<float>  m_Energy;
   vector<double> m_Life;
   vector<double> m_Time;
